requirements.txt
libusb
pyusb